Extract common parts for schedule generation
Moves the common parts of `BookableSlotGenerator` and
`BookableSlotRegenerator` into a shared module to reduce duplication.
